git-svn-id: svn://svn.ijw.co.nz/svn/OpenRa@1329 993157c7-ee19-0410-b2c4-bb4e9862e678

+++ OpenRa.Game/Actor.cs
@@ -13,6 +13,6 @@ namespace OpenRa.Game
 		public abstract float2 RenderLocation { get; }
 		public int palette;
 		public abstract Sprite[] CurrentImages { get; }
-		public virtual void Tick(Game game, double t) { }
+		public virtual void Tick(Game game, int t) { }
 	}
 }
